+++ open-sse/services/combo.ts
@@ -98,6 +98,14 @@ import type {
   HistoricalLatencyStatsEntry,
 } from "./combo/types.ts";
 
+import {
+  MAX_RR_COUNTERS,
+  rrCounters,
+  rrStickyTargets,
+  clampStickyRoundRobinTargetLimit,
+  getStickyRoundRobinStartIndex,
+  recordStickyRoundRobinSuccess,
+} from "./combo/rrState.ts";
 import { validateResponseQuality, toRetryAfterDisplayValue } from "./combo/validateQuality.ts";
 import {
   TRANSIENT_FOR_SEMAPHORE,
@@ -185,14 +193,7 @@ type QuotaFetchCacheConfig = {
 };
 type ResetWindowConfig = ReturnType<typeof resolveResetWindowConfig>;
 
-// In-memory atomic counter per combo for round-robin distribution
-// Resets on server restart (by design — no stale state)
-// Eviction limits to prevent unbounded memory growth
-const MAX_RR_COUNTERS = 500;
 const MAX_RESET_AWARE_CACHE = 200;
-
-const rrCounters = new Map<string, number>();
-const rrStickyTargets = new Map<string, { executionKey: string; successCount: number }>();
 
 const resetAwareConnectionCache = new Map<
   string,
@@ -202,50 +203,6 @@ const resetAwareQuotaCache = new Map<
   string,
   { fetchedAt: number; quota: unknown; refreshPromise: Promise<unknown> | null }
 >();
-
-function clampStickyRoundRobinTargetLimit(value: unknown): number {
-  const numericValue = Number(value);
-  if (!Number.isFinite(numericValue)) return 1;
-  return Math.min(Math.max(Math.floor(numericValue), 1), 1000);
-}
-
-function getStickyRoundRobinStartIndex(
-  comboName: string,
-  targets: ResolvedComboTarget[],
-  stickyLimit: number
-): { startIndex: number; counter: number } {
-  const sticky = rrStickyTargets.get(comboName);
-  const stickyIndex = sticky
-    ? targets.findIndex((target) => target.executionKey === sticky.executionKey)
-    : -1;
-  if (stickyLimit > 1 && sticky && stickyIndex >= 0 && sticky.successCount < stickyLimit) {
-    return { startIndex: stickyIndex, counter: rrCounters.get(comboName) || 0 };
-  }
-
-  const counter = rrCounters.get(comboName) || 0;
-  return { startIndex: counter % targets.length, counter };
-}
-
-function recordStickyRoundRobinSuccess(
-  comboName: string,
-  target: ResolvedComboTarget,
-  stickyLimit: number,
-  targets: ResolvedComboTarget[]
-): void {
-  const sticky = rrStickyTargets.get(comboName);
-  const successCount = sticky?.executionKey === target.executionKey ? sticky.successCount + 1 : 1;
-  if (successCount >= stickyLimit) {
-    const servedIndex = targets.findIndex((entry) => entry.executionKey === target.executionKey);
-    rrCounters.set(
-      comboName,
-      servedIndex >= 0 ? servedIndex + 1 : (rrCounters.get(comboName) || 0) + 1
-    );
-    rrStickyTargets.delete(comboName);
-    return;
-  }
-
-  rrStickyTargets.set(comboName, { executionKey: target.executionKey, successCount });
-}
 
 function finiteNumberOrNull(value: unknown): number | null {
   const numericValue = Number(value);
